backend_requirements.txt
fastapi
uvicorn[standard]
torch==2.2.2+cpu
torchvision==0.17.2+cpu
pillow
fsspec
anyio
google-cloud-secret-manager
numpy==1.26.4
pytorch-lightning==2.2.4
python-multipart
gcsfs

--extra-index-url https://download.pytorch.org/whl/cpu
